+++ ios-search-ads-sample/AppDelegate.swift
@@ -6,7 +6,6 @@
 //
 
 import UIKit
-import iAd
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
@@ -15,28 +14,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [NSObject: AnyObject]?) -> Bool {
         // Override point for customization after application launch.
-        
-        // Iterate through request attribution details
-        //
-        ADClient.shared().requestAttributionDetails {
-            (attributeDetails:[NSObject:AnyObject]?, error: NSError?) in
-            //
-            if error == nil {
-                // process attribute details
-                // v3
-                // TBD: Details on the spec
-                if let v3: AnyObject = attributeDetails?["Version3.1"] {
-                    if let attributionDict = v3 as? [String : AnyObject]{
-                        // iterate through protocol
-                        if let searchAttribution: AnyObject = attributionDict["iad-attribution"] {
-                            // Keep the attribution result local to this sample.
-                            _ = searchAttribution
-                        }
-                    }
-                }
-            }
-        }
-
         return true
     }
 }